pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spice_gen"
version = "0.1.0"
description = "Automatic SPICE netlist generator from YAML/JSON cell topology descriptions"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
spice_gen = "spice_gen.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/spice_gen"]

[tool.pytest.ini_options]
testpaths = ["tests"]
